requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.12
filelock==3.18.0
flake8==7.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
fonttools==4.57.0
fsspec==2025.3.2
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
locust==2.34.1
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
psutil==7.0.0
py-cpuinfo==9.0.0
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.105
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
